pyproject.toml
# Build System Definition
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Project Metadata
[project]
name = "fl-tesi"
version = "1.0.0"
description = "Federated Learning with PyTorch and Flower for a Computer Science Thesis"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.22.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

# Build Target Configuration
[tool.hatch.build.targets.wheel]
packages = ["."]

# Flower App Configuration
[tool.flwr.app]
publisher = "samuelelolli"

# Links the server and client applications to their respective Python files.
[tool.flwr.app.components]
serverapp = "server_app:app"
clientapp = "client_app:app"

# The configuration section is intentionally removed, ensuring that run_experiments.py
# is the single source of truth for experiment configuration via JSON.

# Flower Federation Profiles
[tool.flwr.federations]
default = "local-sim" # The default profile if none is specified.

# Profile for running simulations on CPU only.
[tool.flwr.federations.local-sim]
options.num-supernodes = 10
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0.0

# Profile for running simulations utilizing GPUs.
[tool.flwr.federations.local-sim-gpu]
options.num-supernodes = 10
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0.5